Update radio bot folder
