+++ radio-bot/.env.example
@@ -1,0 +1,1 @@
+TOKEN=
