+++ radio-bot/.eslintrc.json
@@ -1,6 +1,6 @@
 {
 	"root": true,
-	"extends": "marine/prettier/node",
+	"extends": ["neon/common", "neon/node", "neon/typescript", "neon/prettier"],
 	"parserOptions": {
 		"project": "./tsconfig.eslint.json"
 	},
