+++ radio-bot/.gitignore
@@ -1,3 +1,3 @@
-config.json
+.env
 tsconfig.tsbuildinfo
 dist
