--- radio-bot/config.example.json
@@ -1,6 +1,0 @@
-{
-	"token": "discord_bot_token",
-	"device": "audio_hw_device_id",
-	"type": "pulse",
-	"maxTransmissionGap": 5000
-}
