+++ radio-bot/package.json
@@ -2,13 +2,14 @@
 	"name": "discord-radio-bot",
 	"version": "1.0.0",
 	"description": "A proof-of-concept radio bot for @discordjs/voice",
-	"main": "dist/bot.js",
+	"main": "dist/index.js",
+	"type": "module",
 	"scripts": {
-		"start": "npm run build && node dist/bot.js",
+		"build": "rimraf dist && tsc",
+		"build:check": "tsc --noEmit --incremental false",
 		"lint": "prettier --cache --check . && eslint src --ext mjs,js,ts --cache",
 		"format": "prettier --cache --write . && eslint src --ext mjs,js,ts --fix --cache",
-		"build": "rimraf dist && tsc",
-		"build:check": "tsc --noEmit --incremental false"
+		"start": "npm run build && node --env-file=.env dist/index.js"
 	},
 	"keywords": [
 		"discord",
@@ -22,21 +23,17 @@
 	"author": "Amish Shah <amish@shah.gg>",
 	"license": "UNLICENSE",
 	"dependencies": {
-		"@discordjs/voice": "^0.11.0",
-		"discord.js": "^13.10.3",
-		"libsodium-wrappers": "^0.7.10",
+		"@discordjs/voice": "^0.17.0",
+		"discord.js": "^14.15.3",
+		"libsodium-wrappers": "^0.7.13",
 		"prism-media": "^2.0.0-alpha.0"
 	},
 	"devDependencies": {
-		"@typescript-eslint/eslint-plugin": "^5.36.1",
-		"@typescript-eslint/parser": "^5.36.1",
-		"eslint": "^8.23.0",
-		"eslint-config-marine": "^9.4.1",
-		"eslint-config-prettier": "^8.5.0",
-		"eslint-import-resolver-typescript": "^3.5.0",
-		"eslint-plugin-import": "^2.26.0",
-		"prettier": "^2.7.1",
-		"rimraf": "^3.0.2",
-		"typescript": "^4.8.2"
+		"@types/node": "^20.14.7",
+		"eslint": "^8.57.0",
+		"eslint-config-neon": "^0.1.62",
+		"prettier": "^3.3.2",
+		"rimraf": "^5.0.7",
+		"typescript": "5.4.5"
 	}
 }
