+++ radio-bot/tsconfig.json
@@ -1,4 +1,5 @@
 {
+	"$schema": "https://json.schemastore.org/tsconfig.json",
 	// Mapped from https://www.typescriptlang.org/tsconfig
 	"compilerOptions": {
 		// Type Checking
@@ -8,39 +9,46 @@
 		"noFallthroughCasesInSwitch": true,
 		"noImplicitOverride": true,
 		"noImplicitReturns": true,
+		"noPropertyAccessFromIndexSignature": false,
+		"noUncheckedIndexedAccess": true,
 		"noUnusedLocals": true,
 		"noUnusedParameters": true,
 		"strict": true,
-		"useUnknownInCatchVariables": true,
-		"noUncheckedIndexedAccess": true,
-		"skipLibCheck": true,
 
 		// Modules
-		"module": "CommonJS",
-		"moduleResolution": "node",
+		"allowArbitraryExtensions": false,
+		"allowImportingTsExtensions": false,
+		"module": "ESNext",
+		"moduleResolution": "Bundler",
 		"resolveJsonModule": true,
+		"resolvePackageJsonExports": true,
+		"resolvePackageJsonImports": true,
 
 		// Emit
 		"declaration": true,
-		"importHelpers": true,
-		"importsNotUsedAsValues": "error",
-		"inlineSources": true,
+		"declarationMap": true,
+		"importHelpers": false,
 		"newLine": "lf",
 		"noEmitHelpers": true,
 		"outDir": "dist",
-		"preserveConstEnums": true,
 		"removeComments": false,
 		"sourceMap": true,
-		"esModuleInterop": true,
+
+		// Interop Constraints
+		"allowSyntheticDefaultImports": true,
+		"esModuleInterop": false,
 		"forceConsistentCasingInFileNames": true,
+		"isolatedModules": true,
 
 		// Language and Environment
-		"emitDecoratorMetadata": true,
 		"experimentalDecorators": true,
-		"lib": ["ESNext"],
-		"target": "ES2021",
-		"useDefineForClassFields": true
+		"lib": ["ESNext", "esnext.disposable"],
+		"target": "ESNext",
+		"useDefineForClassFields": true,
+
+		// Completeness
+		"skipLibCheck": true
 	},
-	"include": ["src/*.ts"],
-	"exclude": []
+	"include": ["src/**/*.ts"],
+	"exclude": ["node_modules"]
 }
